requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.9.1
    # via django
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
black==25.1.0
    # via -r requirements.in
click==8.2.1
    # via black
django==5.2.5
    # via
    #   -r requirements.in
    #   django-filter
    #   django-simple-history
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-spectacular
    #   model-bakery
django-filter==25.1
    # via -r requirements.in
django-simple-history==3.10.1
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-nested-routers==0.94.2
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
flake8==7.3.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via -r requirements.in
jsonschema==4.25.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
mccabe==0.7.0
    # via flake8
model-bakery==1.20.5
    # via -r requirements.in
mypy==1.17.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==25.0
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via
    #   black
    #   mypy
pillow==11.3.0
    # via -r requirements.in
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
pycodestyle==2.14.0
    # via flake8
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.11.1
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.3
    # via django
typing-extensions==4.14.1
    # via mypy
uritemplate==4.2.0
    # via drf-spectacular
